rename CloudStorageSecretsModal to DataConnectorSecretsModal
